+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * EOF errors will now be handled on connection attempts. (an EOF error will be raised when a connection has
   been established but gets closed immediately)
 * Command execution on TeamSpeak 3 servers (TeamSpeak3::Server#execute) has been refactored.
+* Add `TeamSpeak3::Server#prepare_command` to build a command that can be executed safely (internally used by
+  `TeamSpeak3::Server#execute`)
 
 ## Release 0.0.2 (2018-03-05)